% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Test ODBC stuff}
\usage{
hello(server, database, mykey)
}
\arguments{
\item{server}{name of server}

\item{database}{name of DB}

\item{mykey}{name of key stored via keyring with permissions to DB}
}
\description{
Test ODBC stuff
}
